fix get_repository
