+++ modules/invenio-resourcesyncclient/invenio_resourcesyncclient/static/js/invenio_resourcesyncclient/resync_client.js
@@ -20,7 +20,7 @@ const urlGetLogs = window.location.origin + "/admin/resync/get_logs";
 const urlSync = window.location.origin + "/admin/resync/run_sync";
 const urltoggleRunning = window.location.origin + "/admin/resync/toggle_auto";
 const urlGetTreeList = window.location.origin + "/api/tree";
-const urlGetRepositoryList = window.location.origin + "/resync/get_repository";
+const urlGetRepositoryList = window.location.origin + "/admin/resync/get_repository";
 const status = JSON.parse($("#status").text())
 const resync_mode = JSON.parse($("#resync_mode").text())
 const saving_format = JSON.parse($("#saving_format").text())
